\name{GammaQ}
\alias{GammaQ}
\title{Helper function for \code{\link{PsiHSig}}}
\usage{
  GammaQ(qq, rho)
}
\arguments{
  \item{qq}{(numeric vector) A convenient variable (see
  Charles Hogg's HTML notes for elaboration).}

  \item{rho}{(numeric vector) The experimental noise at
  each datapoint, in units of the expected average signal
  magnitude.}
}
\description{
  This combination of variables appears frequently in
  calculations for the gradient and Hessian.
}

