\name{new_dataset.character}
\alias{new_dataset.character}
\title{New dataset from datafile}
\usage{
  new_dataset.character(f.name, sigma = NA)
}
\arguments{
  \item{f.name}{(character) The name of a datafile which
  has 2 columns (3 if sigma is included)}

  \item{sigma}{(numeric) Gaussian uncertainty (optionally
  different for each datapoint)}
}
\description{
  New dataset from datafile
}

