\name{new_dataset.numeric}
\alias{new_dataset.numeric}
\title{New dataset from numeric data}
\usage{
  new_dataset.numeric(x, y, sigma)
}
\arguments{
  \item{x}{(numeric vector) X-values where we have data}

  \item{y}{(numeric vector) Measured datapoints}

  \item{sigma}{(numeric) Gaussian uncertainty (optionally
  different for each datapoint)}
}
\description{
  New dataset from numeric data
}

