\name{points.background}
\alias{points.background}
\title{Add a background object to a plot}
\usage{
  points.background(x, x.vals = X(x), ...)
}
\arguments{
  \item{x}{\code{\link{background}} object to plot}

  \item{x.vals}{(numeric vector) x-values to evaluate the
  background function}
}
\description{
  Add a background object to a plot
}

