\name{Sigma}
\alias{Sigma}
\alias{Sigma<-}
\title{Pointwise experimental uncertainty}
\usage{
  Sigma(data)

  Sigma(data, value) <- value
}
\arguments{
  \item{data}{(\code{\link{dataset}} object) The measured
  data}
}
\value{
  (numeric vector) The pointwise experimental uncertainty
  (assuming Gaussian noise)
}
\description{
  Pointwise experimental uncertainty
}

