\name{X.background}
\alias{X.background}
\alias{X<-.background}
\title{Spline knot positions for background function}
\usage{
  X.background(b)

  X.background(b, value) <- value
}
\arguments{
  \item{value}{(numeric vector) New X-values for the
  background spline knots}
}
\description{
  Spline knot positions for background function
}

