\name{XFractions}
\alias{XFractions}
\alias{XFractions<-}
\title{Spline knot positions as fractions of available space}
\usage{
  XFractions(b)

  XFractions(b, value) <- value
}
\arguments{
  \item{b}{(\code{\link{background}} object) An estimated
  background (and associated hyperparameters)}

  \item{value}{(numeric vector) New fractions to}
}
\description{
  Spline knot positions as fractions of available space
}

